refactor(healths): polish health code struct and simplify the logic flow
