+++ internal/api/v1/healths/parser.go
@@ -1,0 +1,114 @@
+package healths
+
+import (
+	"fmt"
+
+	"github.com/bigstack-oss/cube-cos-api/internal/api/query"
+	"github.com/bigstack-oss/cube-cos-api/internal/cubecos"
+	v1 "github.com/bigstack-oss/cube-cos-api/internal/definition/v1"
+)
+
+func (h *helper) parseParamsByHandler() error {
+	switch h.handler {
+	case "getHealthSummary":
+		return h.parseSummaryParams()
+	case "genServiceHealthHistory":
+		return h.parseServiceHealthParams()
+	case "getModuleHealthHistory":
+		return h.parseModuleHealthParams()
+	case "forceRepairModule":
+		return h.parseModuleRepairParams()
+	}
+
+	return nil
+}
+
+func (h *helper) parseSummaryParams() error {
+	var err error
+	h.watch, err = query.GetWatch(h.c)
+	if err != nil {
+		return err
+	}
+
+	h.past, err = query.GetPast(h.c)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (h *helper) parseServiceHealthParams() error {
+	var err error
+	h.watch, err = query.GetWatch(h.c)
+	if err != nil {
+		return err
+	}
+
+	h.past, err = query.GetPast(h.c)
+	if err != nil {
+		return err
+	}
+
+	h.period, err = query.GetPeriod(h.c)
+	if err != nil {
+		return err
+	}
+
+	h.serviceType = h.c.Param("serviceType")
+	if !cubecos.IsValidService(h.serviceType) {
+		return fmt.Errorf("invalid serviceType: %s", h.serviceType)
+	}
+
+	return nil
+}
+
+func (h *helper) parseModuleHealthParams() error {
+	var err error
+	h.watch, err = query.GetWatch(h.c)
+	if err != nil {
+		return err
+	}
+
+	h.past, err = query.GetPast(h.c)
+	if err != nil {
+		return err
+	}
+
+	h.period, err = query.GetPeriod(h.c)
+	if err != nil {
+		return err
+	}
+
+	h.serviceType = h.c.Param("serviceType")
+	if !cubecos.IsValidService(h.serviceType) {
+		return fmt.Errorf("invalid serviceType: %s", h.serviceType)
+	}
+
+	h.moduleType = h.c.Param("moduleType")
+	if !cubecos.IsValidServiceAndModule(h.serviceType, h.moduleType) {
+		return fmt.Errorf("invalid serviceType' %s' or module '%s'", h.serviceType, h.moduleType)
+	}
+
+	return nil
+}
+
+func (h *helper) parseModuleRepairParams() error {
+	var err error
+	h.module, err = h.parseModule()
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (h *helper) parseModule() (*v1.Module, error) {
+	m := h.c.Param("moduleType")
+	module, found := cubecos.Modules[m]
+	if !found {
+		return nil, fmt.Errorf("module(%s) not found", m)
+	}
+
+	return &module, nil
+}
