--- internal/api/v1/healths/repair.go
@@ -1,11 +1,0 @@
-package healths
-
-func (h *helper) requestForceRepair() {
-	req := genForceRepairReq(*h.module)
-	reqQueue.Add(req)
-}
-
-func (h *helper) requestCheckRepair() {
-	req := genCheckRepairReq()
-	reqQueue.Add(req)
-}
